Working on filtering and step resposnse

+++ anklebotOS/include/maxonmotor.h
@@ -15,7 +15,7 @@ extern volatile uint32_t *debugBuffer;
 /* ---- Prototypes ---- */
 void motorInit(void);
 void motorCleanUp(void);
-void motorSetDuty(int16_t duty, volatile int16_t *motorDuty);
+void motorSetDuty(int16_t duty, volatile int32_t *motorDuty);
 void motorEnable(void);
 void motorDisable(void);
 void motorSetDir(uint8_t dorsiflex);
